sta-chatbot/test-requirements.txt
langchain==0.2.11
typing-inspect==0.9.0
typing_extensions==4.11.0
langchain-community==0.2.10
langchain-text-splitters==0.2.2
databricks-vectorsearch==0.39
pypdf==4.2.0
mlflow==2.17.0
pyspark==3.5.3
pytest==7.1.2